refactor: rename LoggerProvider to LoggerSynchronizer
