+++ frontend/src/App.tsx
@@ -18,7 +18,7 @@ import SuspenseBoundary from '@web/components/SuspenseBoundary/SuspenseBoundary'
 import WaitingRoomSkeleton from '@pages/WaitingRoom/WaitingRoom.skeleton';
 import MeetingRoomSkeleton from '@pages/MeetingRoom/MeetingRoom.skeleton';
 import SessionProvider from '@Context/SessionProvider/session';
-import LoggerSynchronizer from '@Context/LoggerProvider';
+import LoggerSynchronizer from '@Context/LoggerSynchronizer';
 import ErrorBoundary from './components/ErrorBoundary';
 import EnvGuard from './components/EnvGuard';
 import { ErrorPage } from './pages/ErrorBoundary';
