+++ frontend/src/Context/LoggerSynchronizer/LoggerSynchronizer.tsx
