+++ frontend/src/Context/LoggerSynchronizer/index.tsx
